requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pytz==2023.3
schedule==1.2.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
